refactor: rename component package from 'component' to 'basic_component'
