+++ design_system/src/main/java/com/cairosquad/design_system/basic_component/BottomSheet.kt
@@ -1,4 +1,4 @@
-package com.cairosquad.design_system.component
+package com.cairosquad.design_system.basic_component
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
