+++ ui/src/main/java/com/cairosquad/ui/search/content/SearchFailContent.kt
@@ -13,9 +13,9 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import com.cairosquad.design_system.R
-import com.cairosquad.design_system.component.InputField
-import com.cairosquad.ui.movio_component.StateMessage
+import com.cairosquad.design_system.basic_component.InputField
 import com.cairosquad.design_system.theme.Theme
+import com.cairosquad.ui.movio_component.StateMessage
 import com.cairosquad.viewmodel.exception.ErrorStatus
 import com.cairosquad.viewmodel.search.SearchInteractionListener
 import com.cairosquad.viewmodel.search.SearchScreenState
