+++ ui/src/main/java/com/cairosquad/ui/search/content/SearchResultContent.kt
@@ -30,13 +30,13 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import com.cairosquad.design_system.R
-import com.cairosquad.ui.movio_component.ArtistCard
-import com.cairosquad.design_system.component.InputField
-import com.cairosquad.ui.movio_component.MovieCard
-import com.cairosquad.ui.movio_component.StateMessage
-import com.cairosquad.design_system.component.TopBar
+import com.cairosquad.design_system.basic_component.InputField
+import com.cairosquad.design_system.basic_component.TopBar
 import com.cairosquad.design_system.text_style.defaultTextStyle
 import com.cairosquad.design_system.theme.Theme
+import com.cairosquad.ui.movio_component.ArtistCard
+import com.cairosquad.ui.movio_component.MovieCard
+import com.cairosquad.ui.movio_component.StateMessage
 import com.cairosquad.viewmodel.search.SearchInteractionListener
 import com.cairosquad.viewmodel.search.SearchScreenState
